Change repository types to Long

+++ server/src/main/java/nl/tudelft/oopp/server/services/BuildingService.java
@@ -38,7 +38,7 @@ public class BuildingService {
      * @param id we search for
      * @return the building with that exact id if it exists or null if not
      */
-    public Optional<Building> getBuilding(Integer id) {
+    public Optional<Building> getBuilding(Long id) {
         return buildingRepository.findById(id);
     }
 
@@ -57,7 +57,7 @@ public class BuildingService {
      * @param id       new one to be updated to
      * @param building to be updated
      */
-    public void updateBuilding(Integer id, Building building) {
+    public void updateBuilding(Long id, Building building) {
         buildingRepository.save(building);
     }
 
@@ -66,7 +66,7 @@ public class BuildingService {
      *
      * @param id to be deleted from the list of buildings
      */
-    public void delete(Integer id) {
+    public void delete(Long id) {
         buildingRepository.deleteById(id);
     }
 }
